Add first version of Formula Powering
Formula powering, intended for use with
small powers, precomputes a formula for the power
and uses the gpu to implement this formula.

This is a rough version that works on the CPU
(via KernelAbstractions' cpu backend).
In the future, I hope to add support for
at least the CUDA and Metal backends.
There also ought to be some possible
optimizations with shared memory and/or
shuffle operations.